@@ -1,9 +1,6 @@
 ﻿using System;
 
-namespace CustomIndexerSample
+public record Person(string FirstName, string LastName, DateTime Birthday)
 {
-    public record Person(string FirstName, string LastName, DateTime Birthday)
-    {
-        public override string ToString() => $"{FirstName} {LastName}";
-    }
+    public override string ToString() => $"{FirstName} {LastName}";
 }
