@@ -1,13 +1,7 @@
-﻿namespace OperatorsSample
+﻿public readonly struct Point
 {
-    public readonly struct Point
-    {
-        public Point(int x, int y)
-        {
-            X = x;
-            Y = y;
-        }
-        public int X { get; }
-        public int Y { get; }
-    }
+    public Point(int x, int y) => (X, Y) = (x, y);
+
+    public int X { get; }
+    public int Y { get; }
 }
